Transitions evaluated hierarchically.
Transitions should be matched up the tree when evaluating a macrostep;
if the current state does not match, its parents might.

The pingpong test was actually incorrect, in that the parent used a
targetless transition to initialize. An explicit init state is needed;
if the parent does it as a transition without target with an always true
condition, that will just init forever.

Alternatively, pingpong could have added a condition to only init if not
already initialized, but an init state is probably better as it won't get
checked on each macrostep in each substate.